Adjust the new readme / demo example a bit

+++ demo/main_demo.rs
@@ -422,14 +422,8 @@ fn main() {
   assert_eq!(V.reversed().drain(0..1), [16.0]);
   static VV: StaticVec<f64, 0> = sortedstaticvec!(f64, []);
   assert_eq!(VV, []);
-  let filled = StaticVec::<StaticVec<f64, 8>, 128>::filled_with_by_index(|i| {
-    staticvec![
-      (i + 1) as f64,
-      (i + 2) as f64,
-      (i + 3) as f64,
-      (i + 4) as f64
-    ]
-    .intersperse(((i + 4) * 4) as f64)
+  let filled = StaticVec::<StaticVec<usize, 8>, 128>::filled_with_by_index(|i| {
+    staticvec![i + 1, i + 2, i + 3, i + 4,].intersperse((i + 4) * 4)
   });
   println!("{:?}", filled);
 }
